Make it more obvious how to get functionality in a spec
I was surprised when a straight copy-paste of the Usage with the suggested Gemfile line didn't actually work. The `require` is necessary. I think it would also work to add a `minitest-implicit-subject.rb` file to the top level of `lib`, which just does `require 'minitest/implicit/subject'`, but I'm not sure if you'd rather do that or just make the instructions a bit more obvious.